refactor: reorganize parser files and introduce new utility functions for argument and redirection handling
