+++ include/parser.h
@@ -6,7 +6,7 @@
 /*   By: tsargsya <tsargsya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/30 14:53:04 by tsargsya          #+#    #+#             */
-/*   Updated: 2025/06/10 18:13:54 by tsargsya         ###   ########.fr       */
+/*   Updated: 2025/06/10 23:37:46 by tsargsya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,13 +48,13 @@ typedef struct s_redir
 	struct s_redir			*next;
 }							t_redir;
 
-// parser.c
-t_cmd						*parse_tokens(t_token *tokens, t_shell *sh);
-void						free_cmd_list(t_cmd *cmd);
-
-// redir_utils.c
-void						add_redirection(t_cmd *cmd, t_redir_type type,
-								const char *filename, int quoted);
+// cmd_args_builder.c
+t_cmd						*init_cmd(void);
+int							expand_args(t_cmd *cmd);
+int							append_arg(t_cmd *cmd, char *arg);
+char						*build_argument(t_token **tokens, t_shell *sh);
+char						*build_heredoc_delim(t_token **tokens, int *quoted,
+								t_shell *sh);
 
 // expand_vars.c
 char						*expand_vars(const char *input, t_shell *sh);
@@ -64,4 +64,23 @@ void						append_char_to_result(const char *input,
 								size_t *pos, char **result);
 int							process_dollar_sequence(const char *input,
 								size_t *pos, char **result, t_shell *sh);
+
+// free_utils.c
+void						free_redirs(t_redir *redir);
+void						free_cmd_list(t_cmd *cmd);
+
+// parser.c
+t_cmd						*parse_tokens(t_token *tokens, t_shell *sh);
+
+// redir_utils.c
+void						add_redirection(t_cmd *cmd, t_redir_type type,
+								const char *filename, int quoted);
+void						apply_redirection(t_cmd *cmd, int redirect_type,
+								const char *value, int quoted);
+// token_utils.c
+int							is_arg_token(t_token_type type);
+int							is_redir_token(t_token_type type);
+int							cmd_is_valid_before_pipe(t_cmd *cmd);
+char						*get_chunk_for_token(t_token *tok, t_shell *sh);
+
 #endif
