+++ src/parser/free_utils.c
@@ -1,0 +1,58 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   free_utils.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tsargsya <tsargsya@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/06/10 22:39:27 by tsargsya          #+#    #+#             */
+/*   Updated: 2025/06/10 22:48:12 by tsargsya         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "ft_printf.h"
+#include "libft.h"
+#include "parser.h"
+#include <stddef.h>
+#include <stdio.h>
+#include <stdlib.h>
+
+void	free_redirs(t_redir *redir)
+{
+	t_redir	*next;
+
+	while (redir)
+	{
+		next = redir->next;
+		free(redir->filename);
+		free(redir);
+		redir = next;
+	}
+}
+
+void	free_cmd_list(t_cmd *cmd)
+{
+	t_cmd	*next;
+	int		i;
+
+	while (cmd)
+	{
+		next = cmd->next;
+		if (cmd->args)
+		{
+			i = 0;
+			while (cmd->args[i])
+			{
+				free(cmd->args[i]);
+				i++;
+			}
+			free(cmd->args);
+		}
+		if (cmd->in_redirs)
+			free_redirs(cmd->in_redirs);
+		if (cmd->out_redirs)
+			free_redirs(cmd->out_redirs);
+		free(cmd);
+		cmd = next;
+	}
+}
