+++ src/parser/redir_utils.c
@@ -6,7 +6,7 @@
 /*   By: tsargsya <tsargsya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/23 10:19:57 by tsargsya          #+#    #+#             */
-/*   Updated: 2025/06/10 16:21:44 by tsargsya         ###   ########.fr       */
+/*   Updated: 2025/06/10 22:46:42 by tsargsya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -67,4 +67,17 @@ void	add_redirection(t_cmd *cmd, t_redir_type type, const char *filename,
 		add_redir_node(&cmd->in_redirs, redir);
 	else if (type == REDIR_OUT || type == REDIR_APPEND)
 		add_redir_node(&cmd->out_redirs, redir);
+}
+
+void	apply_redirection(t_cmd *cmd, int redirect_type, const char *value,
+		int quoted)
+{
+	if (redirect_type == TOK_LESS)
+		add_redirection(cmd, REDIR_IN, value, 0);
+	else if (redirect_type == TOK_GREATER)
+		add_redirection(cmd, REDIR_OUT, value, 0);
+	else if (redirect_type == TOK_DLESS)
+		add_redirection(cmd, REDIR_HEREDOC, value, quoted);
+	else if (redirect_type == TOK_DGREATER)
+		add_redirection(cmd, REDIR_APPEND, value, 0);
 }
