+++ src/parser/token_utils.c
@@ -1,0 +1,54 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   token_utils.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tsargsya <tsargsya@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/06/10 22:42:57 by tsargsya          #+#    #+#             */
+/*   Updated: 2025/06/10 23:33:13 by tsargsya         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "ft_printf.h"
+#include "libft.h"
+#include "parser.h"
+#include <stddef.h>
+#include <stdio.h>
+#include <stdlib.h>
+
+int	is_arg_token(t_token_type type)
+{
+	return (type == TOK_WORD || type == TOK_SQUOTED || type == TOK_DQUOTED);
+}
+
+int	is_redir_token(t_token_type type)
+{
+	return (type == TOK_LESS || type == TOK_GREATER || type == TOK_DLESS
+		|| type == TOK_DGREATER);
+}
+
+int	cmd_is_valid_before_pipe(t_cmd *cmd)
+{
+	if (!cmd)
+		return (0);
+	if (cmd->args && cmd->args[0])
+		return (1);
+	if (cmd->in_redirs || cmd->out_redirs)
+		return (1);
+	return (0);
+}
+
+char	*get_chunk_for_token(t_token *tok, t_shell *sh)
+{
+	char	*chunk;
+
+	chunk = NULL;
+	if (tok->type == TOK_SQUOTED)
+		chunk = ft_strdup(tok->value);
+	else if (tok->type == TOK_WORD || tok->type == TOK_DQUOTED)
+		chunk = expand_vars(tok->value, sh);
+	if (!chunk && (tok->type == TOK_SQUOTED || tok->type == TOK_DQUOTED))
+		chunk = ft_strdup("");
+	return (chunk);
+}
